% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-image.R
\name{ovdata_example_image}
\alias{ovdata_example_image}
\title{Example volleyball images}
\usage{
ovdata_example_image(choice = "190301_kats_beds")
}
\arguments{
\item{choice}{string: one or more files to return
\itemize{
\item "190301_kats_beds" - an image from a match between GKS Katowice and MKS Bedzin during the 2018/19 Polish Plus Liga
}}
}
\value{
A character vector with the file path(s)
}
\description{
Example volleyball images
}
\examples{
myfile <- ovdata_example_image()

}
