% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-roster.R
\name{ovdata_example_roster}
\alias{ovdata_example_roster}
\title{Example volleyball team roster files}
\usage{
ovdata_example_roster(
  choice = "Stadium Mirandola",
  as = "path",
  simplify = TRUE
)
}
\arguments{
\item{choice}{string: one or more files to return
\itemize{
\item "Stadium Mirandola" - the team roster for Stadium Mirandola (file from \url{http://volleynetworkitalia.it/})
}}

\item{as}{string: either "path" (return the path to the file) or "parsed" (parse the file into an R data structure using \code{\link[datavolley:dv_read_sq]{datavolley::dv_read_sq()}}}

\item{simplify}{logical: by default if \code{as} is "parsed", the returned list is the same length as \code{choice} where each entry is a parsed object. If \code{simplify} is \code{TRUE} and we have asked for a single file (i.e. \code{length(choice) == 1}, then return just that object (not a list containing that object)}
}
\value{
If \code{as} is "path", a character vector with the file path(s), otherwise a list
}
\description{
Example volleyball team roster files
}
\examples{
myfile <- ovdata_example_roster()
## read this file with
## x <- datavolley::dv_read_sq(myfile)

## or have it read automatically
x <- ovdata_example_roster(as = "parsed")

}
\seealso{
\code{\link[=ovdata_example]{ovdata_example()}}
}
