% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{dv_read}
\alias{pv_read}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{datavolley}{\code{\link[datavolley]{dv_read}}}

  \item{peranavolley}{\code{\link[peranavolley]{pv_read}}}
}}

